PricingNRevenue/LongTripOpt/requirements.txt
# ============================================================
#  urbanblack-aiml — Python dependencies
#  Install: pip install -r requirements.txt
# ============================================================

# ── Data & ML ────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
imbalanced-learn>=0.12.0      # SMOTE
joblib>=1.3.0

# ── API ──────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── Visualisation ────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Database (LongTripOpt loader) ────────────────────────────
psycopg2-binary>=2.9.9        # PostgreSQL driver

# ── PDF extraction ───────────────────────────────────────────
pdfplumber>=0.11.0

# ── Config ───────────────────────────────────────────────────
pyyaml>=6.0.1
